+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [2.2.0] - 2019-06-28
 
 - New function `uef_num:ctz/1`.
 - New function `uef_num:lsb_pos/1`.